backend/requirements.txt
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
requests==2.24.0
